+++ tests/testthat/test-columns.R
@@ -65,8 +65,10 @@ test_that("colDef", {
 })
 
 test_that("colDef aggregate function", {
-  aggregators <- list(JS("function(values, rows) { return 1 }"),
-                      "mean", "sum", "max", "min", "count", "unique", "frequency")
+  aggregators <- list(
+    JS("function(values, rows) { return 1 }"),
+    "mean", "sum", "max", "min", "median", "count", "unique", "frequency"
+  )
   for (func in aggregators) {
     col <- colDef(aggregate = func)
     expect_equal(col$aggregate, func)