+++ NEWS.md
@@ -10,6 +10,7 @@
 * In `reactable()`, the `rowClass`, `rowStyle`, and `details` JavaScript functions
   now receive a `rowInfo.selected` property indicating whether the row is selected.
 * `reactable()` gains a `language` argument to customize the language strings in a table.
+* `colDef()` gains a `"median"` aggregate function to calculate the median of numbers.
 
 ### Breaking changes